selectors refactor, load from lib
