+++ mttl/models/containers/selectors/__init__.py
@@ -1,0 +1,4 @@
+from mttl.models.containers.selectors.base_selectors import *
+from mttl.models.containers.selectors.phatgoose_selector import PhatgooseSelector, PhatgooseSelectorConfig
+from mttl.models.containers.selectors.arrow_selector import ArrowSelector, ArrowSelectorConfig
+from mttl.models.containers.selectors.average_activation_selector import AverageActivationSelector, AverageActivationSelectorConfig
