+++ tests/test_expert_model.py
@@ -6,7 +6,7 @@ from transformers import AutoModelForCausalLM
 from mttl.models.expert_model import MultiExpertModel
 from mttl.models.containers import get_modules_to_modify_trie
 from mttl.models.modifiers.lora import LoRAConfig
-from mttl.models.containers.selectors import (
+from mttl.models.containers.selectors.selectors import (
     PolySelector,
     PolySelectorConfig,
     TaskNameSelectorConfig,
